Checkstyle: Fix up issues in datagen (#1857)
* Remove datagen checkstyle suppressions and refactor code.
