+++ ksql-examples/src/main/java/io/confluent/ksql/datagen/DataGen.java
@@ -16,17 +16,18 @@
 
 package io.confluent.ksql.datagen;
 
+import com.google.common.collect.ImmutableMap;
 import io.confluent.avro.random.generator.Generator;
-import io.confluent.ksql.util.KsqlConfig;
 import java.io.FileInputStream;
 import java.io.IOException;
 import java.io.InputStream;
 import java.util.Arrays;
-import java.util.Collections;
+import java.util.Map;
 import java.util.Objects;
 import java.util.Optional;
 import java.util.Properties;
 import java.util.Random;
+import java.util.function.BiConsumer;
 
 // CHECKSTYLE_RULES.OFF: ClassDataAbstractionCoupling
 public final class DataGen {
@@ -59,7 +60,8 @@ public final class DataGen {
     }
 
     final Generator generator = new Generator(arguments.schemaFile, new Random());
-    final DataGenProducer dataProducer = getProducer(arguments);
+    final DataGenProducer dataProducer = new ProducerFactory()
+        .getProducer(arguments.format, arguments.schemaRegistryUrl);
     final Properties props = getProperties(arguments);
 
     dataProducer.populateTopic(
@@ -82,30 +84,6 @@ public final class DataGen {
     }
 
     return props;
-  }
-
-  private static DataGenProducer getProducer(final Arguments arguments) {
-    switch (arguments.format) {
-      case AVRO:
-        return new AvroProducer(
-            new KsqlConfig(
-                Collections.singletonMap(
-                    KsqlConfig.SCHEMA_REGISTRY_URL_PROPERTY,
-                    arguments.schemaRegistryUrl
-                )
-            )
-        );
-
-      case JSON:
-        return new JsonProducer();
-
-      case DELIMITED:
-        return new DelimitedProducer();
-
-      default:
-        throw new IllegalArgumentException("Invalid format in '" + arguments.format
-                                           + "'; was expecting one of AVRO, JSON, or DELIMITED%n");
-    }
   }
 
   private static void usage() {
@@ -173,6 +151,22 @@ public final class DataGen {
     }
 
     private static final class Builder {
+
+      private static final Map<String, BiConsumer<Builder, String>> ARG_HANDLERS =
+          ImmutableMap.<String, BiConsumer<Builder, String>>builder()
+              .put("quickstart", (builder, argVal) -> builder.quickstart = parseQuickStart(argVal))
+              .put("bootstrap-server", (builder, argVal) -> builder.bootstrapServer = argVal)
+              .put("schema", (builder, argVal) -> builder.schemaFile = toFileInputStream(argVal))
+              .put("format", (builder, argVal) -> builder.format = parseFormat(argVal))
+              .put("topic", (builder, argVal) -> builder.topicName = argVal)
+              .put("key", (builder, argVal) -> builder.keyName = argVal)
+              .put("iterations", (builder, argVal) -> builder.iterations = parseIterations(argVal))
+              .put("maxInterval",
+                  (builder, argVal) -> builder.maxInterval = parseIterations(argVal))
+              .put("schemaRegistryUrl", (builder, argVal) -> builder.schemaRegistryUrl = argVal)
+              .put("propertiesFile",
+                  (builder, argVal) -> builder.propertiesFile = toFileInputStream(argVal))
+              .build();
 
       private Quickstart quickstart;
 
@@ -313,56 +307,44 @@ public final class DataGen {
           ));
         }
 
-        switch (argName) {
-          case "quickstart":
-            try {
-              quickstart = Quickstart.valueOf(argValue.toUpperCase());
-            } catch (final IllegalArgumentException iae) {
-              throw new ArgumentParseException(String.format(
-                  "Invalid quickstart in '%s'; was expecting one of "
-                  + Arrays.toString(Quickstart.values())
-                  + " (case-insensitive)",
-                  argValue
-              ));
-            }
-            break;
-          case "bootstrap-server":
-            bootstrapServer = argValue;
-            break;
-          case "schema":
-            schemaFile = new FileInputStream(argValue);
-            break;
-          case "format":
-            format = parseFormat(argValue);
-            break;
-          case "topic":
-            topicName = argValue;
-            break;
-          case "key":
-            keyName = argValue;
-            break;
-          case "iterations":
-            iterations = parseIterations(argValue);
-            break;
-          case "maxInterval":
-            maxInterval = parseIterations(argValue);
-            break;
-          case "schemaRegistryUrl":
-            schemaRegistryUrl = argValue;
-            break;
-          case "propertiesFile":
-            propertiesFile = new FileInputStream(argValue);
-            break;
-          default:
-            throw new ArgumentParseException(String.format(
-                "Unknown argument name in '%s'",
-                argName
-            ));
-        }
+        setArg(argName, argValue);
         return this;
       }
 
-      private Format parseFormat(final String formatString) {
+      private void setArg(final String argName, final String argVal) {
+        final BiConsumer<Builder, String> handler = ARG_HANDLERS.get(argName);
+        if (handler == null) {
+          throw new ArgumentParseException(String.format(
+              "Unknown argument name in '%s'",
+              argName
+          ));
+        }
+
+        handler.accept(this, argVal);
+      }
+
+      private static FileInputStream toFileInputStream(final String argVal) {
+        try {
+          return new FileInputStream(argVal);
+        } catch (final Exception e) {
+          throw new IllegalArgumentException("File not found: " + argVal, e);
+        }
+      }
+
+      private static Quickstart parseQuickStart(final String argValue) {
+        try {
+          return Quickstart.valueOf(argValue.toUpperCase());
+        } catch (final IllegalArgumentException iae) {
+          throw new ArgumentParseException(String.format(
+              "Invalid quickstart in '%s'; was expecting one of "
+              + Arrays.toString(Quickstart.values())
+              + " (case-insensitive)",
+              argValue
+          ));
+        }
+      }
+
+      private static Format parseFormat(final String formatString) {
         try {
           return Format.valueOf(formatString.toUpperCase());
         } catch (final IllegalArgumentException exception) {
@@ -374,7 +356,7 @@ public final class DataGen {
         }
       }
 
-      private int parseIterations(final String iterationsString) {
+      private static int parseIterations(final String iterationsString) {
         try {
           final int result = Integer.valueOf(iterationsString, 10);
           if (result <= 0) {
